requirements.txt
requests
pystray
Pillow
